rename raspi4 to raspi4-64
