+++ bsp/raspberry-pi/raspi4-64/Kconfig
@@ -20,7 +20,7 @@ source "$PKGS_DIR/Kconfig"
 
 config BCM2711_SOC
     bool
-    select ARCH_ARM_CORTEX_A72
+    select ARCH_ARMV8
     select RT_USING_COMPONENTS_INIT
     select RT_USING_USER_MAIN
     select ARCH_CPU_64BIT
