+++ bsp/raspberry-pi/raspi4-64/README.md
@@ -11,6 +11,38 @@
 
 ## 2. 编译说明
 
+
+
+### 2.1 Window上的环境搭建
+
+Windows环境下推荐使用[env工具](https://www.rt-thread.org/page/download.html)进行编译。
+
+首先下载windows上的aarch64的gcc交叉编译工具，版本为gcc-arm-8.3选择aarch64-elf就可以。
+
+将推荐将gcc解压到`\env\tools\gnu_gcc\arm_gcc`目录下。
+
+接着修改`bsp\raspberry-pi\raspi4-64\rtconfig.py`
+
+修改路径：
+
+```
+EXEC_PATH = r'E:/env_released_1.1.2/env/tools/gnu_gcc/arm_gcc/gcc-arm-8.3-2019.03-i686-mingw32-aarch64-elf/bin'
+```
+
+然后在`bsp\raspberry-pi\raspi4-64\`下输入scons编译即可。
+
+**window环境搭建注意**
+
+下载完成`gcc-arm-8.3-2019.03-i686-mingw32-aarch64-elf.tar.xz`交叉编译工具链后，最好采用7-zip解压工具进行两次解压。
+确保解压目录下的`/bin/aarch64-elf-ld.exe`文件的size不为0。
+否则编译会出现如下错误:
+
+```
+collect2.exe:fatal error:CreateProcess:No such file or directory
+```
+
+### 2.2 Linux上的环境搭建
+
 Linux下推荐使用[gcc工具][2]。Linux版本下gcc版本可采用`gcc-arm-8.3-2019.03-x86_64-aarch64-elf`。
 
 将工具链解压到指定目录,并修改当前bsp下的`EXEC_PATH`为自定义gcc目录。
@@ -20,7 +52,7 @@ PLATFORM    = 'gcc'
 EXEC_PATH   = r'/opt/gcc-arm-8.3-2019.03-x86_64-aarch64-elf/bin/'  
 ```
 
-直接进入`bsp\raspberry-pi\raspi4`，输入scons编译即可。
+直接进入`bsp\raspberry-pi\raspi4-64`，输入scons编译即可。
 
 
 ## 3. 执行
