+++ bsp/raspberry-pi/raspi4-64/link.lds
