+++ bsp/raspberry-pi/raspi4-64/rtconfig.h
@@ -43,6 +43,7 @@
 #define RT_CONSOLE_DEVICE_NAME "uart"
 #define RT_VER_NUM 0x40003
 #define ARCH_CPU_64BIT
+#define ARCH_ARMV8
 
 /* RT-Thread Components */
 
@@ -153,12 +154,6 @@
 
 
 /* samples: kernel and components samples */
-
-
-/* Privated Packages of RealThread */
-
-
-/* Network Utilities */
 
 #define BCM2711_SOC
